third_party/WebKit/LayoutTests/external/wpt/background-fetch/interfaces-worker.https-expected.txt

This is a testharness.js-based test.
PASS Service Worker-scoped tests.
PASS Exposed interfaces in a Service Worker.
FAIL ServiceWorkerRegistration interface: existence and properties of interface object assert_false: expected false got true
FAIL ServiceWorkerGlobalScope interface: attribute onbackgroundfetched assert_true: The prototype object must have a property "onbackgroundfetched" expected true got false
FAIL ServiceWorkerGlobalScope interface: attribute onbackgroundfetchfail assert_true: The prototype object must have a property "onbackgroundfetchfail" expected true got false
FAIL ServiceWorkerGlobalScope interface: attribute onbackgroundfetchabort assert_true: The prototype object must have a property "onbackgroundfetchabort" expected true got false
FAIL ServiceWorkerGlobalScope interface: attribute onbackgroundfetchclick assert_true: The prototype object must have a property "onbackgroundfetchclick" expected true got false
PASS BackgroundFetchManager interface: existence and properties of interface object
PASS BackgroundFetchManager interface object length
PASS BackgroundFetchManager interface object name
PASS BackgroundFetchManager interface: existence and properties of interface prototype object
PASS BackgroundFetchManager interface: existence and properties of interface prototype object's "constructor" property
PASS BackgroundFetchManager interface: operation fetch(DOMString, [object Object],[object Object], BackgroundFetchOptions)
PASS BackgroundFetchManager interface: operation get(DOMString)
PASS BackgroundFetchManager interface: operation getIds()
FAIL BackgroundFetchRegistration interface: existence and properties of interface object Cannot read property 'has_extended_attribute' of undefined
PASS BackgroundFetchRegistration interface object length
PASS BackgroundFetchRegistration interface object name
FAIL BackgroundFetchRegistration interface: existence and properties of interface prototype object Cannot read property 'has_extended_attribute' of undefined
PASS BackgroundFetchRegistration interface: existence and properties of interface prototype object's "constructor" property
PASS BackgroundFetchRegistration interface: attribute id
PASS BackgroundFetchRegistration interface: attribute uploadTotal
PASS BackgroundFetchRegistration interface: attribute uploaded
PASS BackgroundFetchRegistration interface: attribute downloadTotal
PASS BackgroundFetchRegistration interface: attribute downloaded
FAIL BackgroundFetchRegistration interface: attribute activeFetches assert_true: The prototype object must have a property "activeFetches" expected true got false
PASS BackgroundFetchRegistration interface: attribute onprogress
PASS BackgroundFetchRegistration interface: operation abort()
PASS BackgroundFetchFetch interface: existence and properties of interface object
PASS BackgroundFetchFetch interface object length
PASS BackgroundFetchFetch interface object name
PASS BackgroundFetchFetch interface: existence and properties of interface prototype object
PASS BackgroundFetchFetch interface: existence and properties of interface prototype object's "constructor" property
PASS BackgroundFetchFetch interface: attribute request
FAIL BackgroundFetchActiveFetches interface: existence and properties of interface object assert_own_property: self does not have own property "BackgroundFetchActiveFetches" expected property "BackgroundFetchActiveFetches" missing
FAIL BackgroundFetchActiveFetches interface object length assert_own_property: self does not have own property "BackgroundFetchActiveFetches" expected property "BackgroundFetchActiveFetches" missing
FAIL BackgroundFetchActiveFetches interface object name assert_own_property: self does not have own property "BackgroundFetchActiveFetches" expected property "BackgroundFetchActiveFetches" missing
FAIL BackgroundFetchActiveFetches interface: existence and properties of interface prototype object assert_own_property: self does not have own property "BackgroundFetchActiveFetches" expected property "BackgroundFetchActiveFetches" missing
FAIL BackgroundFetchActiveFetches interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "BackgroundFetchActiveFetches" expected property "BackgroundFetchActiveFetches" missing
FAIL BackgroundFetchActiveFetches interface: operation match(RequestInfo) assert_own_property: self does not have own property "BackgroundFetchActiveFetches" expected property "BackgroundFetchActiveFetches" missing
FAIL BackgroundFetchActiveFetches interface: operation values() assert_own_property: self does not have own property "BackgroundFetchActiveFetches" expected property "BackgroundFetchActiveFetches" missing
FAIL BackgroundFetchActiveFetch interface: existence and properties of interface object assert_own_property: self does not have own property "BackgroundFetchActiveFetch" expected property "BackgroundFetchActiveFetch" missing
FAIL BackgroundFetchActiveFetch interface object length assert_own_property: self does not have own property "BackgroundFetchActiveFetch" expected property "BackgroundFetchActiveFetch" missing
FAIL BackgroundFetchActiveFetch interface object name assert_own_property: self does not have own property "BackgroundFetchActiveFetch" expected property "BackgroundFetchActiveFetch" missing
FAIL BackgroundFetchActiveFetch interface: existence and properties of interface prototype object assert_own_property: self does not have own property "BackgroundFetchActiveFetch" expected property "BackgroundFetchActiveFetch" missing
FAIL BackgroundFetchActiveFetch interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "BackgroundFetchActiveFetch" expected property "BackgroundFetchActiveFetch" missing
FAIL BackgroundFetchActiveFetch interface: attribute responseReady assert_own_property: self does not have own property "BackgroundFetchActiveFetch" expected property "BackgroundFetchActiveFetch" missing
FAIL BackgroundFetchEvent interface: existence and properties of interface object Cannot read property 'has_extended_attribute' of undefined
PASS BackgroundFetchEvent interface object length
PASS BackgroundFetchEvent interface object name
FAIL BackgroundFetchEvent interface: existence and properties of interface prototype object Cannot read property 'has_extended_attribute' of undefined
PASS BackgroundFetchEvent interface: existence and properties of interface prototype object's "constructor" property
PASS BackgroundFetchEvent interface: attribute id
FAIL BackgroundFetchSettledEvent interface: existence and properties of interface object assert_own_property: self does not have own property "BackgroundFetchSettledEvent" expected property "BackgroundFetchSettledEvent" missing
FAIL BackgroundFetchSettledEvent interface object length assert_own_property: self does not have own property "BackgroundFetchSettledEvent" expected property "BackgroundFetchSettledEvent" missing
FAIL BackgroundFetchSettledEvent interface object name assert_own_property: self does not have own property "BackgroundFetchSettledEvent" expected property "BackgroundFetchSettledEvent" missing
FAIL BackgroundFetchSettledEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "BackgroundFetchSettledEvent" expected property "BackgroundFetchSettledEvent" missing
FAIL BackgroundFetchSettledEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "BackgroundFetchSettledEvent" expected property "BackgroundFetchSettledEvent" missing
FAIL BackgroundFetchSettledEvent interface: attribute fetches assert_own_property: self does not have own property "BackgroundFetchSettledEvent" expected property "BackgroundFetchSettledEvent" missing
FAIL BackgroundFetchSettledFetches interface: existence and properties of interface object assert_own_property: self does not have own property "BackgroundFetchSettledFetches" expected property "BackgroundFetchSettledFetches" missing
FAIL BackgroundFetchSettledFetches interface object length assert_own_property: self does not have own property "BackgroundFetchSettledFetches" expected property "BackgroundFetchSettledFetches" missing
FAIL BackgroundFetchSettledFetches interface object name assert_own_property: self does not have own property "BackgroundFetchSettledFetches" expected property "BackgroundFetchSettledFetches" missing
FAIL BackgroundFetchSettledFetches interface: existence and properties of interface prototype object assert_own_property: self does not have own property "BackgroundFetchSettledFetches" expected property "BackgroundFetchSettledFetches" missing
FAIL BackgroundFetchSettledFetches interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "BackgroundFetchSettledFetches" expected property "BackgroundFetchSettledFetches" missing
FAIL BackgroundFetchSettledFetches interface: operation match(RequestInfo) assert_own_property: self does not have own property "BackgroundFetchSettledFetches" expected property "BackgroundFetchSettledFetches" missing
FAIL BackgroundFetchSettledFetches interface: operation values() assert_own_property: self does not have own property "BackgroundFetchSettledFetches" expected property "BackgroundFetchSettledFetches" missing
PASS BackgroundFetchSettledFetch interface: existence and properties of interface object
FAIL BackgroundFetchSettledFetch interface object length assert_equals: wrong value for BackgroundFetchSettledFetch.length expected 0 but got 2
PASS BackgroundFetchSettledFetch interface object name
PASS BackgroundFetchSettledFetch interface: existence and properties of interface prototype object
PASS BackgroundFetchSettledFetch interface: existence and properties of interface prototype object's "constructor" property
PASS BackgroundFetchSettledFetch interface: attribute response
FAIL BackgroundFetchUpdateEvent interface: existence and properties of interface object assert_own_property: self does not have own property "BackgroundFetchUpdateEvent" expected property "BackgroundFetchUpdateEvent" missing
FAIL BackgroundFetchUpdateEvent interface object length assert_own_property: self does not have own property "BackgroundFetchUpdateEvent" expected property "BackgroundFetchUpdateEvent" missing
FAIL BackgroundFetchUpdateEvent interface object name assert_own_property: self does not have own property "BackgroundFetchUpdateEvent" expected property "BackgroundFetchUpdateEvent" missing
FAIL BackgroundFetchUpdateEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "BackgroundFetchUpdateEvent" expected property "BackgroundFetchUpdateEvent" missing
FAIL BackgroundFetchUpdateEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "BackgroundFetchUpdateEvent" expected property "BackgroundFetchUpdateEvent" missing
FAIL BackgroundFetchUpdateEvent interface: operation updateUI(DOMString) assert_own_property: self does not have own property "BackgroundFetchUpdateEvent" expected property "BackgroundFetchUpdateEvent" missing
PASS BackgroundFetchClickEvent interface: existence and properties of interface object
PASS BackgroundFetchClickEvent interface object length
PASS BackgroundFetchClickEvent interface object name
FAIL BackgroundFetchClickEvent interface: existence and properties of interface prototype object Cannot read property 'has_stringifier' of undefined
PASS BackgroundFetchClickEvent interface: existence and properties of interface prototype object's "constructor" property
PASS BackgroundFetchClickEvent interface: attribute state
Harness: the test ran to completion.
